pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.18.0",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
]
